Add a vim mode line to not expand tabs into spaces
According to PEP8 using spaces is the preferred over tabs in python. However
this project uses tabs throughout, and thus we feel it is a good idea to
explicitly instruct vim to use tabs (and not spaces) to ensure consistency, as
some people may have their im setup to follow PEP8. It might be nice to have
such a modestring for emacs too.