A new approach using the --contains flag
This algorithm establishes a forward and reverse mapping for all labels, using
the 'git tag --contains' and 'git branch --contains' commands. The forward
mapping indicates what other labels contain this label. The forward mapping
indicates who the parent labels of a label are.  Then it tries to establish the
parents we want to display by removing stuff from the forward mapping.

Works on simple toy examples, but takes forever and fails on the git via git
repo. :-((