Include the graph reduction algo
We now use the graph reduction algorithm from alg-simple as a pre-processing to
the recursion. I.e. we perform the recursion on the reduced graph, and thereby
(hopefully) avoid reaching the maximum recusrion depth.

Works smoothly for the git via git repo, and is maybe faster than
'_optimize_non_labels()'.