Major bug fix:
Until now _real_ tags have been ignored, and only lightweight ones were included
in the graph. Now we include them. However not the sha1 of the tag is used, but
instead the sha1 that the tag points to.

Still the problem remains, what if tags, point to tags, point to tags?